+++ Middleware/src/main/java/ca/concordia/encs/citydata/producers/GeometryProducer.java
@@ -3,10 +3,9 @@ package ca.concordia.encs.citydata.producers;
 import java.security.InvalidParameterException;
 
 import ca.concordia.encs.citydata.core.contracts.IOperation;
-import ca.concordia.encs.citydata.core.contracts.IProducer;
 import ca.concordia.encs.citydata.core.contracts.IRunner;
-import ca.concordia.encs.citydata.core.implementations.AbstractProducer;
-import ca.concordia.encs.citydata.producers.base.JSONProducer;
+import ca.concordia.encs.citydata.core.implementations.JSONProducer;
+import ca.concordia.encs.citydata.core.utils.RequestOptions;
 
 /**
  * This Producer outputs GeoJSON geometries for a given city.
@@ -15,7 +14,11 @@ import ca.concordia.encs.citydata.producers.base.JSONProducer;
  * @since 2025-05-28
  */
 
-public class GeometryProducer extends AbstractProducer<String> implements IProducer<String> {
+public class GeometryProducer extends JSONProducer {
+	public GeometryProducer(String filePath, RequestOptions fileOptions) {
+		super(filePath, fileOptions);
+	}
+
 	private JSONProducer jsonProducer;
 
 	public void setCity(String city) {
