+++ Middleware/src/test/java/ca/concordia/encs/citydata/test/producers/BaseProducersTest.java
@@ -9,9 +9,9 @@ import java.util.UUID;
 import org.junit.jupiter.api.Test;
 
 import ca.concordia.encs.citydata.core.contracts.IProducer;
+import ca.concordia.encs.citydata.core.implementations.CSVProducer;
+import ca.concordia.encs.citydata.core.implementations.JSONProducer;
 import ca.concordia.encs.citydata.datastores.InMemoryDataStore;
-import ca.concordia.encs.citydata.producers.base.CSVProducer;
-import ca.concordia.encs.citydata.producers.base.JSONProducer;
 import ca.concordia.encs.citydata.runners.SingleStepRunner;
 
 /**
